# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "## TMF API Reference: TMF656 - Service Test Management\n\n### Release\
    \ : 18.5 - December 2018\n\nService Test Management API goal is to provide the\
    \ ability to manage tests of provisioned Services. \n\n### Resource \n- ServiceTest\
    \ \n- ServiceTestSpecification\n\n### Operations\nService Test Management API\
    \ performs the following operations on the resources :\n- Retrieve an entity or\
    \ a collection of entities depending on filter criteria\n- Partial update of an\
    \ entity (including updating rules)\n- Create an entity (including default values\
    \ and creation rules)\n- Delete an entity (for administration purposes)\n- Manage\
    \ notification of events"
  version: "4.0.0"
  title: "Service Test Management"
host: "localhost:8080"
basePath: "/tmf-api"
...
schemes:
- "http"
consumes:
- "application/json;charset=utf-8"
produces:
- "application/json;charset=utf-8"
paths:
  /serviceTest:
    get:
      tags:
      - "serviceTest"
      summary: "List or find ServiceTest objects"
      description: "This operation list or find ServiceTest entities"
      operationId: "listServiceTest"
      parameters:
      - name: "fields"
        in: "query"
        description: "Comma-separated properties to be provided in response"
        required: false
        type: "string"
      - name: "offset"
        in: "query"
        description: "Requested index for start of resources to be provided in response"
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        description: "Requested number of resources to be provided in response"
        required: false
        type: "integer"
      responses:
        200:
          description: "Success"
          headers:
            X-Total-Count:
              type: "integer"
              description: "Total number of items matching criteria"
            X-Result-Count:
              type: "integer"
              description: "Actual number of items returned in the response body"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ServiceTest"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceTest"
    post:
      tags:
      - "serviceTest"
      summary: "Creates a ServiceTest"
      description: "This operation creates a ServiceTest entity."
      operationId: "createServiceTest"
      parameters:
      - in: "body"
        name: "serviceTest"
        description: "The ServiceTest to be created"
        required: true
        schema:
          $ref: "#/definitions/ServiceTest_Create"
      responses:
        201:
          description: "Created"
          schema:
            $ref: "#/definitions/ServiceTest"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceTest"
  /serviceTest/{id}:
    get:
      tags:
      - "serviceTest"
      summary: "Retrieves a ServiceTest by ID"
      description: "This operation retrieves a ServiceTest entity. Attribute selection\
        \ is enabled for all first level attributes."
      operationId: "retrieveServiceTest"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifier of the ServiceTest"
        required: true
        type: "string"
      - name: "fields"
        in: "query"
        description: "Comma-separated properties to provide in response"
        required: false
        type: "string"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/ServiceTest"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceTest"
    delete:
      tags:
      - "serviceTest"
      summary: "Deletes a ServiceTest"
      description: "This operation deletes a ServiceTest entity."
      operationId: "deleteServiceTest"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifier of the ServiceTest"
        required: true
        type: "string"
      responses:
        204:
          description: "Deleted"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceTest"
    patch:
      tags:
      - "serviceTest"
      summary: "Updates partially a ServiceTest"
      description: "This operation updates partially a ServiceTest entity."
      operationId: "patchServiceTest"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifier of the ServiceTest"
        required: true
        type: "string"
      - in: "body"
        name: "serviceTest"
        description: "The ServiceTest to be updated"
        required: true
        schema:
          $ref: "#/definitions/ServiceTest_Update"
      responses:
        200:
          description: "Updated"
          schema:
            $ref: "#/definitions/ServiceTest"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceTest"
...
definitions:
  Any: {}
...
  Characteristic:
    type: "object"
    required:
    - "name"
    - "value"
    properties:
      name:
        type: "string"
        description: "Name of the characteristic"
      valueType:
        type: "string"
        description: "Data type of the value of the characteristic"
      value:
        description: "The value of the characteristic"
        $ref: "#/definitions/Any"
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class"
      '@schemaLocation':
        type: "string"
        format: "uri"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class entity name"
    description: "Describes a given characteristic of an object or entity through\
      \ a name/value pair."
    example:
      '@baseType': "@baseType"
      '@type': "@type"
      valueType: "valueType"
      name: "name"
      '@schemaLocation': "http://example.com/aeiou"
      value: {}
...
  ServiceRef:
    type: "object"
    required:
    - "href"
    - "id"
    properties:
      id:
        type: "string"
        description: "Id of the service"
      href:
        type: "string"
        description: "reference of the service"
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class"
      '@schemaLocation':
        type: "string"
        format: "uri"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class entity name"
      '@referredType':
        type: "string"
        description: "The actual type of the target instance when needed for disambiguation."
    description: "Service reference, for when Service is used by other entities"
    example:
      '@referredType': "@referredType"
      '@baseType': "@baseType"
      '@type': "@type"
      id: "id"
      href: "href"
      '@schemaLocation': "http://example.com/aeiou"
...
  ServiceTest:
    type: "object"
    properties:
      id:
        type: "string"
        description: "The identifier of the service test"
      href:
        type: "string"
        description: "Hyperlink to access the service test"
      description:
        type: "string"
        description: "Description of the service test"
      endDateTime:
        type: "string"
        format: "date-time"
        description: "The end date and time of the service test"
      mode:
        type: "string"
        description: "An indication of whether the service test is running in \n\"\
          PROACTIVE\" or \"ONDEMAND\" mode"
      name:
        type: "string"
        description: "The name of the service test"
      startDateTime:
        type: "string"
        format: "date-time"
        description: "The start date and time of the service test."
      state:
        type: "string"
        description: "The actual state the service test is in"
      characteristic:
        type: "array"
        description: "List of characteristics with values that define the test run"
        items:
          $ref: "#/definitions/Characteristic"
      relatedService:
        description: "The actual service being tested"
        $ref: "#/definitions/ServiceRef"
      testMeasure:
        type: "array"
        description: "The results of the test in terms of the measured metrics"
        items:
          $ref: "#/definitions/TestMeasure"
      testSpecification:
        description: "The specification for this test"
        $ref: "#/definitions/ServiceTestSpecificationRef"
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class"
      '@schemaLocation':
        type: "string"
        format: "uri"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class entity name"
    description: "A service test is an entity that exists for a controlled test invocation\
      \ on a service. The service \ntest is executed according to a schedule and contains\
      \ service test configuration parameters that are to be \napplied at execution\
      \ time, and service test measures that result."
    example:
      '@type': "@type"
      description: "description"
      testMeasure:
      - metricDescription: "metricDescription"
        metricName: "metricName"
        unitOfMeasure: "unitOfMeasure"
        '@baseType': "@baseType"
        captureDateTime: "2000-01-23T04:56:07.000+00:00"
        '@type': "@type"
        accuracy: 0.8008282
        captureMethod: "captureMethod"
        ruleViolation:
        - conformanceComparatorLower: "conformanceComparatorLower"
          conformanceComparatorUpper: "conformanceComparatorUpper"
          '@type': "@type"
          description: "description"
          thresholdRuleSeverity: "thresholdRuleSeverity"
          appliedConsequence:
          - appliedAction: "appliedAction"
            repeatAction: true
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            description: "description"
            '@schemaLocation': "http://example.com/aeiou"
          - appliedAction: "appliedAction"
            repeatAction: true
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            description: "description"
            '@schemaLocation': "http://example.com/aeiou"
          '@baseType': "@baseType"
          tolerancePeriod:
            amount: ""
            '@baseType': "@baseType"
            '@type': "@type"
            units: "units"
            '@schemaLocation': "http://example.com/aeiou"
          numberOfAllowedCrossing: 6
          conformanceTargetUpper: "conformanceTargetUpper"
          name: "name"
          '@schemaLocation': "http://example.com/aeiou"
          conformanceTargetLower: "conformanceTargetLower"
        - conformanceComparatorLower: "conformanceComparatorLower"
          conformanceComparatorUpper: "conformanceComparatorUpper"
          '@type': "@type"
          description: "description"
          thresholdRuleSeverity: "thresholdRuleSeverity"
          appliedConsequence:
          - appliedAction: "appliedAction"
            repeatAction: true
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            description: "description"
            '@schemaLocation': "http://example.com/aeiou"
          - appliedAction: "appliedAction"
            repeatAction: true
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            description: "description"
            '@schemaLocation': "http://example.com/aeiou"
          '@baseType': "@baseType"
          tolerancePeriod:
            amount: ""
            '@baseType': "@baseType"
            '@type': "@type"
            units: "units"
            '@schemaLocation': "http://example.com/aeiou"
          numberOfAllowedCrossing: 6
          conformanceTargetUpper: "conformanceTargetUpper"
          name: "name"
          '@schemaLocation': "http://example.com/aeiou"
          conformanceTargetLower: "conformanceTargetLower"
        '@schemaLocation': "http://example.com/aeiou"
        metricHref: "metricHref"
        value:
          '@baseType': "@baseType"
          '@type': "@type"
          valueType: "valueType"
          name: "name"
          '@schemaLocation': "http://example.com/aeiou"
          value: {}
      - metricDescription: "metricDescription"
        metricName: "metricName"
        unitOfMeasure: "unitOfMeasure"
        '@baseType': "@baseType"
        captureDateTime: "2000-01-23T04:56:07.000+00:00"
        '@type': "@type"
        accuracy: 0.8008282
        captureMethod: "captureMethod"
        ruleViolation:
        - conformanceComparatorLower: "conformanceComparatorLower"
          conformanceComparatorUpper: "conformanceComparatorUpper"
          '@type': "@type"
          description: "description"
          thresholdRuleSeverity: "thresholdRuleSeverity"
          appliedConsequence:
          - appliedAction: "appliedAction"
            repeatAction: true
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            description: "description"
            '@schemaLocation': "http://example.com/aeiou"
          - appliedAction: "appliedAction"
            repeatAction: true
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            description: "description"
            '@schemaLocation': "http://example.com/aeiou"
          '@baseType': "@baseType"
          tolerancePeriod:
            amount: ""
            '@baseType': "@baseType"
            '@type': "@type"
            units: "units"
            '@schemaLocation': "http://example.com/aeiou"
          numberOfAllowedCrossing: 6
          conformanceTargetUpper: "conformanceTargetUpper"
          name: "name"
          '@schemaLocation': "http://example.com/aeiou"
          conformanceTargetLower: "conformanceTargetLower"
        - conformanceComparatorLower: "conformanceComparatorLower"
          conformanceComparatorUpper: "conformanceComparatorUpper"
          '@type': "@type"
          description: "description"
          thresholdRuleSeverity: "thresholdRuleSeverity"
          appliedConsequence:
          - appliedAction: "appliedAction"
            repeatAction: true
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            description: "description"
            '@schemaLocation': "http://example.com/aeiou"
          - appliedAction: "appliedAction"
            repeatAction: true
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            description: "description"
            '@schemaLocation': "http://example.com/aeiou"
          '@baseType': "@baseType"
          tolerancePeriod:
            amount: ""
            '@baseType': "@baseType"
            '@type': "@type"
            units: "units"
            '@schemaLocation': "http://example.com/aeiou"
          numberOfAllowedCrossing: 6
          conformanceTargetUpper: "conformanceTargetUpper"
          name: "name"
          '@schemaLocation': "http://example.com/aeiou"
          conformanceTargetLower: "conformanceTargetLower"
        '@schemaLocation': "http://example.com/aeiou"
        metricHref: "metricHref"
        value:
          '@baseType': "@baseType"
          '@type': "@type"
          valueType: "valueType"
          name: "name"
          '@schemaLocation': "http://example.com/aeiou"
          value: {}
      endDateTime: "2000-01-23T04:56:07.000+00:00"
      relatedService:
        '@referredType': "@referredType"
        '@baseType': "@baseType"
        '@type': "@type"
        id: "id"
        href: "href"
        '@schemaLocation': "http://example.com/aeiou"
      characteristic:
      - '@baseType': "@baseType"
        '@type': "@type"
        valueType: "valueType"
        name: "name"
        '@schemaLocation': "http://example.com/aeiou"
        value: {}
      - '@baseType': "@baseType"
        '@type': "@type"
        valueType: "valueType"
        name: "name"
        '@schemaLocation': "http://example.com/aeiou"
        value: {}
      testSpecification:
        '@referredType': "@referredType"
        '@baseType': "@baseType"
        '@type': "@type"
        id: "id"
        href: "href"
        '@schemaLocation': "http://example.com/aeiou"
        version: "version"
      mode: "mode"
      startDateTime: "2000-01-23T04:56:07.000+00:00"
      '@baseType': "@baseType"
      name: "name"
      id: "id"
      href: "href"
      state: "state"
      '@schemaLocation': "http://example.com/aeiou"
...
  ServiceTestSpecificationRef:
    type: "object"
    required:
    - "href"
    - "id"
    properties:
      id:
        type: "string"
        description: "Identifier of a service test specification."
      href:
        type: "string"
        description: "Hyperlink to access a service test specification."
      version:
        type: "string"
        description: "Version of a service test specification "
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class"
      '@schemaLocation':
        type: "string"
        format: "uri"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class entity name"
      '@referredType':
        type: "string"
        description: "The actual type of the target instance when needed for disambiguation."
    description: "The service test specification used by the service test."
    example:
      '@referredType': "@referredType"
      '@baseType': "@baseType"
      '@type': "@type"
      id: "id"
      href: "href"
      '@schemaLocation': "http://example.com/aeiou"
      version: "version"
...
  Error:
    type: "object"
    required:
    - "code"
    - "reason"
    properties:
      code:
        type: "integer"
        description: "Application relevant detail, defined in the API or a common\
          \ list."
      reason:
        type: "string"
        description: "Explanation of the reason for the error which can be shown to\
          \ a client user."
      message:
        type: "string"
        description: "More details and corrective actions related to the error which\
          \ can be shown to a client user."
      status:
        type: "integer"
        description: "HTTP Error code extension"
      referenceError:
        type: "string"
        format: "uri"
        description: "URI of documentation describing the error."
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class."
      '@schemaLocation':
        type: "string"
        format: "uri"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class entity name."
    description: "Used when an API throws an Error, typically with a HTTP error response-code\
      \ (3xx, 4xx, 5xx)"